Added some assets and messed around with positioning
Added some assets and messed around with the positioning. The base functionality of the program itself is ready. As soon as the user clicks on listen the browser mic is ready. What's mostly needed now is just tweaking to the pixel art itself and properly implementing it into the website.